Moved some duplicate code into a helper file
