+++ Jantimizer/ExperimentSuite/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using DatabaseConnector.Exceptions;
+using ExperimentSuite.Helpers;
 using ExperimentSuite.Models;
 using ExperimentSuite.Models.ExperimentParsing;
 using ExperimentSuite.SuiteDatas;
@@ -42,22 +43,9 @@ namespace ExperimentSuite
         public MainWindow()
         {
             InitializeComponent();
-            var iconHandle = ExperimentSuite.Properties.Resources.icon;
-            this.Icon = ByteToImage(iconHandle);
+            var iconHandle = Properties.Resources.icon;
+            this.Icon = ImageHelper.ByteToImage(iconHandle);
             CacheViewerControl.Toggle(true);
-        }
-
-        public static ImageSource ByteToImage(byte[] imageData)
-        {
-            BitmapImage biImg = new BitmapImage();
-            MemoryStream ms = new MemoryStream(imageData);
-            biImg.BeginInit();
-            biImg.StreamSource = ms;
-            biImg.EndInit();
-
-            ImageSource imgSrc = biImg as ImageSource;
-
-            return imgSrc;
         }
 
         private async Task RunExperiments()
